Add virt-run to run batches on a clean vm.
This script will clone an existing domain, start the cloned guest, run a
series of commands, then delete the clone. The use case is to run builds
and tests on a pristine guest on each build cycle.

Commands starting with '@' are reserved as local commands. The '@reboot'
command will reboot the clone and wait for the reboot to complete.

Example usage:

    virt-run \
    --base ubuntu1610-builder \
    --user buildbot \
    "uname -a" \
    "kernel-ppa get" \
    "kernel-ppa install" \
    "@reboot" \
    "uname -a" \
    "git clone git://git.myrepo.org/myproject" \
    "cd myproject && ./bootstrap.sh && ./configure && make" \
    "cd myproject && sudo make install" \
    "run-tests"