kernel-ppa: dont list bogus names for versions
List only versions which start with the letter 'v', so we do not
accidently get the wrong thing when trying to get the most recent
version. This change means we do not list some very old releases on
kernel.ubuntu.com, but we only care about leading edge versions.

Before this change:

    $ kernel-ppa list
    ?C=N;O=A
    /~kernel-ppa
    v4.17-rc4
    v4.17-rc3
    v4.17-rc2
    ...

After this change:

    $ kernel-ppa list
    v4.17-rc4
    v4.17-rc3
    v4.17-rc2
    v4.17-rc1
    v4.16
    ...